+++ app/services/landable/render_service.rb
@@ -18,6 +18,7 @@ module Landable
       content = render_template(page.body, {}, registers: {
         page: page,
         assets: assets_for_page,
+        responder: options[:responder],
       })
 
       if layout?
@@ -73,11 +74,11 @@ module Landable
       @assets_for_theme ||= theme ? theme.assets_as_hash : {}
     end
 
-    def render_template template, variables = nil, options = nil
-      options ||= {}
+    def render_template template, variables = nil, liquid_options = nil
+      liquid_options ||= {}
       variables ||= {}
 
-      parse(template).render!(variables, options)
+      parse(template).render!(variables, liquid_options)
     end
   end
 end