+++ spec/dummy/config/initializers/landable.rb
@@ -6,6 +6,7 @@ Landable.configure do |config|
   config.sitemap_exclude_categories = %w(Testing)
   config.sitemap_protocol = 'https'
   config.sitemap_additional_paths = %w(/terms.html)
+  config.partials_to_templates = %w(partials/foobazz)
 
   config.reserved_paths = %w(/reserved_path_set_in_initializer)
 end