+++ spec/factories/template.rb
@@ -4,5 +4,10 @@ FactoryGirl.define do
     description "Factory-generated template"
     thumbnail_url "http://example.com/bogus-screenshot.png"
     body '<div class="container">content goes here!</div>'
+
+    trait :partial do
+      editable false
+      file     'random/partial'
+    end
   end
 end